Drop support for system or user installed packaging
As of now, system or user installed packaging is imported first.
This was made to make it possible for distributions to debundle
packaging from this project. But in practice, it can lead to
unpredictable results and can't be properly supported.

`pyproject-installer` is supposed to not have any external dependencies
and should bring everything it needs inside itself (bundle).

Fixes: https://github.com/stanislavlevin/pyproject_installer/issues/62
Signed-off-by: Stanislav Levin <slev@altlinux.org>